bak/requirements.txt
numpy
pandas
pyyaml
tensorflow
pytest
keras
asyncio
websockets
termcolor
crossplane